+++ src/PHPSandbox/SandboxWhitelistVisitor.php
@@ -13,7 +13,7 @@ namespace PHPSandbox;
  * @namespace PHPSandbox
  *
  * @author  Elijah Horton <fieryprophet@yahoo.com>
- * @version 1.2
+ * @version 1.3
  */
 class SandboxWhitelistVisitor extends \PHPParser_NodeVisitorAbstract {
     /** The PHPSandbox instance to check against