+++ toolkit/index.php
@@ -119,6 +119,7 @@
             'blacklist' => $blacklist,
             'definitions' => $definitions
         ));
+        $sandbox->set_error_handler(function(\PHPSandbox\Error $e){ echo '<h2 style="color: red;">' . $e->getMessage() . '</h2>'; });
         try {
             ob_start();
             if($setup_code){
@@ -135,8 +136,8 @@
             $buffer = ob_get_contents();
             ob_end_clean();
             die('<pre>' . $buffer . '</pre>');
-        } catch(\PHPSandbox\Error $error){
-            throw $error;
+        } catch(\PHPSandbox\Error $e){
+            echo '<h2 style="color: red;">' . $e->getMessage() . '</h2>';
         }
     }